another attempt at remodelling after removing int parameters

+++ application/models.py
@@ -12,8 +12,8 @@ class Questions(db.Model):
 class Answers(db.Model):
     __tablename__ = "Answers"
     id = db.Column(db.Integer(), primary_key=True)
-    users_answers = db.Column(db.Integer(10), unique=True)
-    correct_answer = db.Column(db.Integer(10), nullable=False)
+    users_answers = db.Column(db.Integer(), unique=True)
+    correct_answer = db.Column(db.Integer(), nullable=False)
     question = relationship("Questions", backref = "answer")
 
     def __repr_(self):
